+++ scripts/pca_full.py
@@ -7,17 +7,14 @@ from dscribe.descriptors import SOAP
 from sklearn.decomposition import PCA
 import joblib
 
-from shared import *
+from tools import *
 
-
-FULL_ATOM = ['POPC_charmm36m_303K_PO4_NC3_C4A_C4B.npz', 
-             'POPC_gromos43a1-s3_303K_PO4_NC3_C4A_C4B.npz',
-             'POPC_lipid17_303K_PO4_NC3_C4A_C4B.npz',
-             'POPC_slipids_303K_PO4_NC3_C4A_C4B.npz']
 
 def main(system, cutoff, sample, overwrite=True):
 
-    prefix = "{}/Lipids/dscribe_{}{}/full_soap/{}_ang/".format(HOME, system, TR, cutoff)
+    prefix = (
+        "{}/Lipids/dscribe_{}{}/full_soap/{}_ang/".format(HOME, system, TR, cutoff)
+    )
 
     X = []
 
@@ -33,7 +30,9 @@ def main(system, cutoff, sample, overwrite=True):
     pca = PCA(n_components=10)
     pca.fit(X)
 
-    folder = "{}/Lipids/dscribe_{}{}/pca/{}_ang/".format(HOME, system, TR, cutoff)
+    folder = (
+        "{}/Lipids/dscribe_{}{}/pca/{}_ang/".format(HOME, system, TR, cutoff)
+    )
 
     if not os.path.isdir(folder):
         os.mkdir(folder)
