+++ scripts/soap_describe.py
@@ -5,13 +5,8 @@ import argparse
 import numpy as np
 from dscribe.descriptors import SOAP
 
-from shared import *
+from tools import *
 
-
-species = {
-    "4beads": ["N", "P", "C"],
-    "phosphate": ["P"]
-}
 
 def get_folder(average):
     return "avg_soap" if average == "inner" else "full_soap"
@@ -19,10 +14,18 @@ def get_folder(average):
 
 def main(system, cutoff, average, overwrite=True):
 
-    files = sorted(glob.glob('{}/Lipids/trajectories_{}{}/POPC_*xyz'.format(HOME, system, TR)))
-    print('Processing trajectories at 303k')
+    files = sorted(
+        glob.glob(
+            "{}/Lipids/trajectories_{}{}/POPC_*xyz"
+            .format(HOME, system, TR)
+        )
+    )
+    print("Processing trajectories at 303k")
     files = [i for i in files if '303' in i]
-    folder = '{}/Lipids/dscribe_{}{}/{}/{}_ang/'.format(HOME, system, TR, get_folder(average), cutoff)
+    folder = (
+        "{}/Lipids/dscribe_{}{}/{}/{}_ang/"
+        .format(HOME, system, TR, get_folder(average), cutoff)
+    )
 
     if not os.path.isdir(folder):
         os.mkdir(folder)
