+++ scripts/tools/shared.py
@@ -1,112 +1,9 @@
-import os
-import itertools
-
 import numpy as np
-from ase.io import read
-from Pipeline.DPA import DensityPeakAdvanced
+# from Pipeline.DPA import DensityPeakAdvanced
 from sklearn.neighbors import NearestNeighbors, KNeighborsClassifier
-from scipy.stats import entropy
 
-
-HOME = os.path.expanduser("~")
-
-TR = "_1us"
-
-MAX_TRAJ_SIZE = 4000
-
-def grouper(iterable, n):
-    """
-    Group iterable exaclty by n.
-    """
-    if len(iterable) % n != 0:
-        raise ValueError(f"Iterable's length is not a multiple of {n}")
-    return zip(*(iter(iterable),) * n)
-
-
-def frame2string(coord):
-    """
-    Convert a xyz frame to string.
-    """
-    write = ""
-    for i, frame in enumerate(coord):
-        write += "%d\nAutoGen_%d\n" % (len(coord[0]), i)
-        for mol in frame:
-            write += "       N{:12.5f}{:11.5f}{:11.5f}\n".format(*mol)
-    return write
-
-
-def read_traj(filename, index=":", start=None, end=None, stride=None):
-    """
-    Read xyz trajectory into a ase.Atoms object.
-    """
-    if all([start, end, stride]):
-        index = "{}:{}:{}".format(start, end, stride),
-    return read(filename, index=index, format="xyz")
-
-
-def str2bool(v):
-    """
-    Convert string to boolean for argument parsing. 
-    """
-    if isinstance(v, bool):
-       return v
-    if v.lower() in ('yes', 'true', 't', 'y', '1'):
-        return True
-    elif v.lower() in ('no', 'false', 'f', 'n', '0'):
-        return False
-    else:
-        raise ValueError('Boolean value expected.')
-
-
-def KernelSoap(x, y, n):
-    """
-    Soap Kernel
-    """
-    return ( np.dot(x, y) / (np.dot(x, x) * np.dot(y, y)) ** 0.5 ) ** n
-
-
-def DistanceSoap(x, y, n=1):
-    """
-    Distance based on Soap Kernel.
-    """
-    try:
-        return (2.0 - 2.0 * KernelSoap(x, y, n)) ** 0.5
-    except FloatingPointError:
-        return 0
-
-
-def cartesian_product(arrays):
-    """
-    Explicit grid genereation.
-    """
-    la = len(arrays)
-    dtype = np.result_type(*arrays)
-    arr = np.empty([la] + [len(a) for a in arrays], dtype=dtype)
-    for i, a in enumerate(np.ix_(*arrays)):
-        arr[i, ...] = a
-    return arr.reshape(la, -1).T
-
-
-def _lazy_cartesian_product(ranges):
-    """
-    Generate grid using iteratools.product.
-    """
-    return itertools.product(*ranges)
-
-
-def lazy_cartesian_product(arrays, size=100):
-    """
-    Group generator in array of size `size`.
-    """
-    x = []
-    for i,v in enumerate(_lazy_cartesian_product(arrays)):
-        if not (i + 1) % size:
-            x.append(v)
-            yield np.array(x)
-            x = []
-        else:
-            x.append(v)
-    yield np.array(x)
+from .utils import cartesian_product, lazy_cartesian_product
+from .metrics import JS
 
 
 class UniformGrid:
@@ -139,22 +36,9 @@ class UniformGrid:
             self.minmax[dim] = self.get_ranges(x[:, dim])
         return self
     
-    def transform(self, n):
-        ranges = []
-        for dim in range(self.dim):
-            spacing = ( self.minmax[dim][1] - self.minmax[dim][0] ) / n
-            ranges.append(
-                np.linspace(
-                    self.minmax[dim][0] - spacing / 2, 
-                    self.minmax[dim][1] + spacing / 2, 
-                    n
-                )
-            )
-        return cartesian_product(ranges)
-
-    def lazy_transform(self, n, chunk=100):
+    def _transform(self, n):
         """
-        Create grid as a generator in chunk is size `chunk` for lazy evaluation.
+        Returns list of ranges for grid creation.
         """
         ranges = []
         for dim in range(self.dim):
@@ -165,10 +49,23 @@ class UniformGrid:
                         self.minmax[dim][0] - spacing / 2, 
                         self.minmax[dim][1] + spacing / 2, 
                         n
-                )
+                    )
                 )
             )
-        return lazy_cartesian_product(ranges, chunk)
+        return ranges
+
+    def transform(self, n):
+        """
+        Create grid as a generator in chunk is size `chunk` for lazy evaluation.
+        """
+        return cartesian_product(self._transform(n))
+
+
+    def lazy_transform(self, n, chunk=100):
+        """
+        Create grid as a generator in chunk is size `chunk` for lazy evaluation.
+        """
+        return lazy_cartesian_product(self._transform(n), chunk)
     
 
 def fit_grid_refiner(grid, sample_points, neigh=10):
@@ -214,21 +111,6 @@ def extract_sample(files, sample_size=5000):
     return np.vstack(X)
 
 
-def KL(p, q):
-    """
-    Kullback-Leibler divergence
-    """
-    return np.sum(p * np.log(p / q))
-
-
-def JS(p, q):
-    """
-    Jensen–Shannon divergence
-    """
-    m = (p + q) / 2
-    return (entropy(p, m)[0] + entropy(q, m)[0]) / 2
-
-
 def predict(d, k, x):
     """
     Use NearestNeighbors to interpolate densities on out of sample values.
@@ -269,5 +151,4 @@ def calculate_js(p, files, fine_grid, n, size):
         kl = JS(p, q)
         kls.append(kl)
         print(file, kl)
-        
     return kls
